+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 All notable changes to the "coffee-status" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
+
+## [0.2.0] - 2019-Apr 30
+
+### Added
+
+    -Timer
+    -Counter reset
+        -CommandPallet: Reset Coffee Count
 
 ## [0.1.3] - 2019-Apr 17